refactor(benchmark): rewrite scalars and points gen to preprocess
