+++ mopro-core/Cargo.toml
@@ -58,6 +58,9 @@ duration-string = { version = "0.0.6", optional = true }
 rand = { version = "0.8.0", optional = true }
 rand_chacha = { version = "0.3.1", optional = true }
 lazy_static = { version = "1.4.0", optional = true }
+ark-ff = { version = "=0.4.1", default-features = false, features = [
+    "parallel",
+] }
 
 # GPU explorations from mopro/Cargo.toml patch
 ark-ec-3 = { git = 'https://github.com/arkworks-rs/algebra.git', package = 'ark-ec', tag = 'v0.3.0', features = ["parallel"], optional = true}
