+++ mopro-core/src/middleware/gpu_explorations/mod.rs
@@ -1,2 +1,3 @@
 pub mod arkworks_pippenger;
 pub mod trapdoortech_zprize_msm;
+pub mod utils;
