+++ mopro-core/src/middleware/mod.rs
@@ -1,4 +1,4 @@
 pub mod circom;
 
-#[cfg(feature = "gpu-benchmarks")]
+// #[cfg(feature = "gpu-benchmarks")]
 pub mod gpu_explorations;
